coordinator/requirements.txt
asyncssh
aiohttp
openrgb-python
